Update root README.md
Using the path `data/2019/2019-01-08` looks in the current directory for the folder data. In contrast, using the path `/data/2019/2019-01-08` looks in the **root** folder for the data folder. 

It would be more consistent if we write our links from the root folder throughout this repo, and this would help (admittedly in a very small way) with the yearly migration of READMEs.